fix(kernel): satisfy the clippy gates on owned packages
The owned-package clippy gates had not executed since the CI front door
broke, accumulating diagnostics across the EEVDF, DMA, and world work.
Mechanical fixes throughout; narrow allows carry the structural
result_large_err, large_enum_variant, and intentional host-test Arc
cases.

+++ kernel/src/syscall/task/clone.rs
@@ -894,7 +894,7 @@ pub fn sys_vfork(caller_memory: UserMemoryCapability, uctx: &UserContext) -> AxR
     sys_clone(
         caller_memory,
         uctx,
-        (CLONE_VM | CLONE_VFORK | SIGCHLD) as u32,
+        CLONE_VM | CLONE_VFORK | SIGCHLD,
         0,
         0,
         0,
